requirements.txt
requests
python-decouple
jsbeautifier
slack_sdk
google-generativeai
python-dotenv
